fix: Renamed configs to config & updated cmd/root.go
